flow: Enable `exact_by_default`.
This was a much-anticipated feature [1] that's available in recent
versions of Flow.

We'll still be able to make object types inexact, but we'll have to
do it explicitly, with a marked syntax. That seems appropriate:
since we don't want to use them as often, it makes sense for them to
be marked, and for exact object types to be unmarked.

As explained in the docs [2],

````
Set this to `true` to indicate that Flow should interpret
object types as exact by default. When this flag is `false`, Flow
has the following behavior:

```js
{foo: number} // inexact
{| foo: number |} // exact
{foo: number, ...} // inexact
```

When this flag is `true`, Flow has the following behavior:

```js
{foo: number} // exact
{| foo: number |} // exact
{foo: number, ...} // inexact
```

The default value is `false`.
````

[1] https://medium.com/flow-type/on-the-roadmap-exact-objects-by-default-16b72933c5cf
[2] https://flow.org/en/docs/config/options/#toc-exact-by-default-boolean