flow: Use implicitly *exact* object types, now that we can!
In the previous commit, we enabled the `exact_by_default`
option [1].

Now, we can make our exact object types unmarked. Since we expect to
use exact object types more often than inexact ones, it makes sense
for the exact ones to be unmarked, and to require the inexact ones
to be marked (with `, ...` as in `{foo: number, ...}`).

To confirm that this works, try adding a `nonsense="123"` prop to
the `Card` callsite in frontend/Components/VaccineInfo/Cards.jsx.

[1] https://flow.org/en/docs/config/options/#toc-exact-by-default-boolean